% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{em.distance.m}
\alias{em.distance.m}
\title{distance between grids or sites}
\usage{
em.distance.m(
  dfcoor,
  between = "site",
  proj = "+proj=longlat",
  transform = "+init=epsg:32754"
)
}
\arguments{
\item{dfcoor}{-- a df with lat and lon, and location id}

\item{between}{-- the location id to calculate between distances}

\item{proj}{-- projection of your lat lons}

\item{transform}{-- projection to transform into (UTM)}
}
\value{
df of distances between 'populations'
}
\description{
simple function to create a dataframe of distance in metres
between sites
}
\examples{
em.distance.m(grids, between = "gridId")
}
\author{
Emily Stringer
}
