% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_functions.R
\name{em.filtering.gl}
\alias{em.filtering.gl}
\title{My Filtering protocol}
\usage{
em.filtering.gl(glx, t.Loc = 0.95, t.Ind = 0.95, t.maf = 0.005)
}
\arguments{
\item{glx}{-- a genlight object}

\item{t.Loc}{-- callrate threshold for loci}

\item{t.Ind}{-- callrate threshold for individuals}

\item{t.maf}{-- threshold for minor allele frequency}
}
\value{
my filtered genlight object
}
\description{
Currently I am filtering on call rates, monomorphs,
minor allele frequency, and secondaries.
}
\examples{
gl <- em.filtering.gl(gl)
}
\author{
Emily Stringer
}
