% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_functions.R
\name{em.filtering.history}
\alias{em.filtering.history}
\title{My Filtering history}
\usage{
em.filtering.history(glx)
}
\arguments{
\item{glx}{-- a filtered genlight object}
}
\value{
filtering history as a df
}
\description{
Shows how I filtered my data in a nice df (usable in kable and DT)
}
\examples{
em.filtering.history(gl)
}
\author{
Emily Stringer
}
