% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstatistics_functions.R
\name{em.fst.trip}
\alias{em.fst.trip}
\title{Within trip fst calculations}
\usage{
em.fst.trip(glx, as.pop = "gridId", min.n = 4, equal = F)
}
\arguments{
\item{glx}{-- a genlight object}

\item{as.pop}{-- defined population to base fst calculations on}

\item{min.n}{-- minimum sample size}

\item{equal}{-- equal sample size}
}
\value{
bootstraps df from gl.fst.pop
}
\description{
fst bootstraps df calculated for a single genlight object.
fst separated by trip and based on your defined populations.
I have also included minimum sample size. You no longer need to
}
\examples{
fst <- em.fst.trip(gl, gridId, 4)
}
\author{
Emily Stringer
}
