% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl_transformation_functions.R
\name{em.gl.assignPop}
\alias{em.gl.assignPop}
\title{reassigne population}
\usage{
em.gl.assignPop(glx, define.pop = "gridId")
}
\arguments{
\item{glx}{-- genlight object}

\item{define.pop}{-- "variable" from ind.metrics to assign as new pop}
}
\value{
genlight object with new pop() values
}
\description{
I created this function to help me with reassigning pop()
for a list of gl objects, using lapply.
}
\examples{
phyear <- seppop(gl) \%>\% 
lapply(., function(x) em.gl.assignPop(x, define.pop = "pop"))
}
\author{
Emily Stringer
}
