% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl_transformation_functions.R
\name{em.gl.min}
\alias{em.gl.min}
\title{minimum population size}
\usage{
em.gl.min(glx, define.pop = "gridId", s.size = 4, equal = FALSE)
}
\arguments{
\item{glx}{-- genlight object}

\item{define.pop}{-- "variable" from ind.metrics to assign as pop}

\item{s.size}{-- minimum sample size}

\item{equal}{-- if each pop should have the same number of samples}
}
\value{
genlight object
}
\description{
removes populations that have less than defined sample size
}
\examples{
em.gl.min(pherm, "gridId", 4, FALSE)
}
\author{
Emily Stringer
}
