% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl_transformation_functions.R
\name{em.gl.remove.singlepop}
\alias{em.gl.remove.singlepop}
\title{remove single population trips}
\usage{
em.gl.remove.singlepop(glx, define.pop = "gridId")
}
\arguments{
\item{glx}{-- genlight object}

\item{define.pop}{-- "variable" from ind.metrics to assign as pop}
}
\value{
genlight object with removed trips
}
\description{
removes trips that only have one pop(), where fst's can't be computed
}
\examples{
phMinSample <- lapply(1:5, function(x) em.gl.min.pop.n(pherm, 
                                define.pop = iNewPop, s.size = x))
}
\author{
Emily Stringer
}
