% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstatistics_functions.R
\name{em.heterozygosity.ind}
\alias{em.heterozygosity.ind}
\title{Individual heterozygosity}
\usage{
em.heterozygosity.ind(gl)
}
\arguments{
\item{gl}{-- a genlight object}
}
\value{
df with ho and gl ind.metrics
}
\description{
seppop based on ind.names and calculates gl.ho
}
\examples{
 heterozygosity <- em.heterozygosity.ind(gl)
}
\author{
Emily Stringer
}
