% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstatistics_functions.R
\name{em.heterozygosity.trip}
\alias{em.heterozygosity.trip}
\title{Within trip fis calculations}
\usage{
em.heterozygosity.trip(glx, as.pop = "gridId", min.n = 4, equal = F)
}
\arguments{
\item{glx}{-- a genlight object}

\item{as.pop}{-- defined population to base fis calculations on}

\item{min.n}{-- minimum sample size}

\item{equal}{-- equal sample size}
}
\value{
df with he, ho, and fis
}
\description{
fis bootstraps df calculated for a single genlight object.
fis separated by trip and based on your defined populations.
}
\examples{
 heterozygosity <- em.heterozygosity.trip(gl)
}
\author{
Emily Stringer
}
