% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/since_event_functions.R
\name{em.months.since}
\alias{em.months.since}
\title{Time sincePeak}
\usage{
em.months.since(eventdate, df)
}
\arguments{
\item{eventdate}{-- date of event}

\item{df}{-- df with variable trip}
}
\value{
df with updated sincePeak and peakNo
}
\description{
This function takes dates and sets the time since peak to zero at
those dates and then every month after is +1. It can now be used on
any dataframe with a 'trip' column and date format yyyy-mm-dd
}
\examples{
em.months.since(c("2004-02-01", "2010-01-01"), df)
}
\author{
Emily Stringer
}
