% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_functions.R
\name{em.setup.repo}
\alias{em.setup.repo}
\title{setup repository}
\usage{
em.setup.repo(
  pushFolders = "code",
  ignoreFolder = c("data", "figures", "output")
)
}
\arguments{
\item{ignoreFolder}{\itemize{
\item folders whose contents will not be pushed to github
}}

\item{pushFolder}{\itemize{
\item folders that will be pushed to github
}}
}
\value{
creates folders and .gitignore files
}
\description{
Setup the folders and git ignore files at the beginning of a repo.
Folder structure is so
}
\details{
\itemize{
\item code (pushed)
\item data (ignored)
\item figures (ignored)
\item output (ignored)
}
}
\examples{
em.setup.repo()

### make function in emphd !! ###

}
\author{
Emily Stringer
}
