% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{em.sim.boombust}
\alias{em.sim.boombust}
\title{conceptual simulation gl.offspring}
\usage{
em.sim.boombust(
  glBase,
  pops = 5,
  popSize = 50,
  dispersal = 0.25,
  dispConstant = FALSE
)
}
\arguments{
\item{glBase}{-- a genlight object to base simulation on (starting snps)}

\item{pops}{-- generations when mixing occurs}

\item{popSize}{-- number of individuals per population during bust}

\item{dispersal}{-- amount of dispersal}
}
\value{
returns genlight for 20 generations
}
\description{
This function creates a simulation using gl.sim.offspring
over 20 generations, where five populations are isolated
except during a few generations when dispersal happens. It
also simulates population increases during across two
generations, the first population increase with dispersal,
the second the next generation after dispersal. I plan on
developing a simulation with constant dispersal as well.
}
\details{
Make sure sex is defined by m and f (only these will be included)
}
\examples{
em.sim.boombust(pherm, pops = 5, popSize = 50, dispersal = 0.25, 
                                                 dispConstant = TRUE)
}
\author{
Emily Stringer
}
