% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{em.sim.wrapper}
\alias{em.sim.wrapper}
\title{wrapper function for em.sim.boombust.ddd}
\usage{
em.sim.wrapper(glBase, subpops = 25, dispersal = 0.1)
}
\arguments{
\item{glBase}{-- a genlight object to base simulation on (starting snps)}

\item{subpops}{-- number of subpopulations}

\item{dispersal}{-- amount of dispersal}
}
\value{
returns genlight for 20 generations
}
\description{
This function simulates the three different types of dispersal on one
genlight object with a select dispersal rate and number of subpopulations.
Population size of subpopulations is set at 20.
}
\examples{
sim <- em.sim.boombust.ddd(gl, supops = 100, dispersal = 0.05)
}
\author{
Emily Stringer
}
