% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstatistics_functions.R
\name{em.stat.combine}
\alias{em.stat.combine}
\title{combines fst or heterozygosities with since event, captures, rainfall,
grid coordinates}
\usage{
em.stat.combine(
  stat,
  eventDates,
  rainCap = NULL,
  grids = NULL,
  as.pop = "gridId"
)
}
\arguments{
\item{stat}{-- fst or heterozygosity df from em.fst.trip}

\item{eventDates}{-- list of dates since event (yyyy-mm-01)}

\item{rainCap}{-- df of rainfall and captures}

\item{grids}{-- grid cordinates df}

\item{as.pop}{-- defined population to base fis calculations on}
}
\value{
combined df
}
\description{
fis bootstraps df calculated for a single genlight object.
fis separated by trip and based on your defined populations.
}
\examples{
 em.stat.combine(heterozygosity, raindates, raincap, grids) 
}
\author{
Emily Stringer
}
