% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{em.sim.boombust.ddd}
\alias{em.sim.boombust.ddd}
\title{conceptual simulation gl.offspring in regard}
\usage{
em.sim.boombust.ddd(glBase, dispersal = 0.1, dispersalType = "positive")
}
\arguments{
\item{glBase}{-- a genlight object to base simulation on (starting snps)}

\item{dispersal}{-- amount of dispersal}

\item{dispersalType}{-- postive, negative, or constant dispersal.}
}
\value{
returns genlight for 20 generations
}
\description{
quick and dirty sim function based on em.sim.boombust.
looking at density-dependent dispersal.
}
\details{
d-d = density dependent.
See em.sim.boombust for more details...
}
\examples{
sim <- em.sim.boombust.ddd(gl, dispersal = 0.05, dispPositive = T)
}
\author{
Emily Stringer
}
