% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{em.sim.boombust.ddd}
\alias{em.sim.boombust.ddd}
\title{conceptual simulation gl.offspring in regard}
\usage{
em.sim.boombust.ddd(
  glBase,
  subpops = 25,
  dispersal = 0.1,
  dispersalType = "positive"
)
}
\arguments{
\item{glBase}{-- a genlight object to base simulation on (starting snps)}

\item{subpops}{-- number of subpopulations}

\item{dispersal}{-- amount of dispersal}

\item{dispersalType}{-- positive, negative, or constant dispersal.}
}
\value{
returns genlight for 20 generations
}
\description{
quick and dirty sim function based on em.sim.boombust.
looking at density-dependent dispersal.
}
\details{
d-d = density dependent.
See em.sim.boombust for more details...
}
\examples{
sim <- em.sim.boombust.ddd(gl)
}
\author{
Emily Stringer
}
